use android studio folder structure
Move these into a folder structure that allows easier integration with Android Studio by matching the package name to the folder tree
